[jsk_recognition_utils] Separate polygon.h from geo_util.h
