+++ jsk_recognition_utils/CMakeLists.txt
@@ -53,6 +53,7 @@ add_library(jsk_recognition_utils SHARED
   src/geo/line.cpp
   src/geo/segment.cpp
   src/geo/plane.cpp
+  src/geo/polygon.cpp
   )
 
 install(TARGETS jsk_recognition_utils
